pyproject.toml
[tool.poetry]
name = "wrt-tech-assignment"
version = "0.1.0"
description = "Web crawling project with Airflow"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
selenium = "^4.15.0"
psycopg2-binary = "^2.9.7"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.0"
apache-airflow-providers-docker = "^3.8.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.0.0"
flake8 = "^6.0.0"
mypy = "^1.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py310']

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true 